+++ internal/types/common.go
@@ -77,6 +77,10 @@ func StripLeadingSlash(folder string) string {
 	return folder
 }
 
+func LocalDockerfile(u *url.URL) (string, error) {
+	return AbsolutePath(u, "Dockerfile")
+}
+
 func AbsolutePath(documentURL *url.URL, path string) (string, error) {
 	documentPath := documentURL.Path
 	if runtime.GOOS == "windows" {